Module ircd::settings

Structs

Args

ircd cli

ChannelInfo

This struct holds information about preconfigured channels. In the TOML configuration file, we can configure channels as such:

ContactInfo

Enums

RPL

Constants

BROADCAST_LAST_TERM_MSG

CONFIG_FILE

CONFIG_FILE_CONTENTS

LIFETIME_FOR_ORPHAN

MAXIMUM_LENGTH_OF_MESSAGE

MAXIMUM_LENGTH_OF_NICKNAME

MAX_CONFIRM

SIZE_OF_IDSS_BUFFER

SIZE_OF_MSGS_BUFFER

TERM_MAX_TIME_DIFFERENCE

TIMEOUT_FOR_RESEND_UNREAD_MSGS

UNREAD_MSG_EXPIRE_TIME

Functions

get_current_time

parse_configured_channels

Parse a TOML string for any configured channels and return a map containing said configurations.

parse_configured_contacts

Parse a TOML string for any configured contact list and return a map containing said configurations.

parse_priv_key 🔒

salt_box_from_shared_secret 🔒